# Conflicts:
#	source/game_sa/Collision/ColBox.h
#	source/game_sa/Collision/ColSphere.h
#	source/game_sa/PlayerInfo.cpp
#	source/game_sa/PlayerInfo.h

+++ source/game_sa/Enums/ePedStats.h
@@ -1,6 +1,6 @@
 #pragma once
 
-enum class ePedStats : uint16 {
+enum class ePedStats : uint32 {
     PLAYER,
     COP,
     MEDIC,
